+++ components/hackathons/registration-form/RegisterFormStep3.tsx
@@ -7,10 +7,21 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
 import { useFormContext } from "react-hook-form";
 import { RegisterFormValues } from "./RegistrationForm";
 import { Checkbox } from "@/components/ui/checkbox";
+
+const hackathonParticipationOptions = [
+  { value: "yes", label: "Yes" },
+  { value: "no", label: "No" },
+];
 import { EventsLang, t } from "@/lib/events/i18n";
 
 interface RegisterFormStep3Props {
@@ -23,7 +34,7 @@ export function RegisterFormStep3({ isOnlineHackathon, lang = "en" }: RegisterFo
 
   return (
     <>
-      {/* Step 3: Terms & Agreements */}
+      {/* Step 2: Terms & Agreements */}
    
       <div className="mb-6">
         <h3 className="text-lg font-semibold text-foreground">{t(lang, "reg.step3.title")}</h3>
@@ -36,7 +47,36 @@ export function RegisterFormStep3({ isOnlineHackathon, lang = "en" }: RegisterFo
         <div className="w-full h-px bg-zinc-300 mt-2" />
       </div>
       <div className="space-y-6">
+        <FormField
+          control={form.control}
+          name="hackathon_participation"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>
+                Have you participated in any other hackathons before? (Optional)
+              </FormLabel>
+              <Select onValueChange={field.onChange} value={field.value}>
+                <FormControl>
+                  <SelectTrigger className="text-zinc-600">
+                    <SelectValue placeholder="Select an option (optional)" />
+                  </SelectTrigger>
+                </FormControl>
+                <SelectContent className="bg-white dark:bg-black border-gray-300 dark:border-zinc-600 text-black dark:text-zinc-600 rounded-md shadow-md">
+                  {hackathonParticipationOptions.map((option) => (
+                    <SelectItem key={option.value} value={option.value}>
+                      {option.label}
+                    </SelectItem>
+                  ))}
+                </SelectContent>
+              </Select>
+              <FormMessage className="text-zinc-600">
+                Let us know if this is your first hackathon or if you have prior experience.
+              </FormMessage>
+            </FormItem>
+          )}
+        />
 
+        <div className="w-full h-px bg-zinc-300" />
 
         <FormField
           control={form.control}
